--- a/.ci/install.sh
+++ b/.ci/install.sh
@@ -3,10 +3,9 @@
 set -ex
 
 if [[ "$(uname -s)" == 'Darwin' ]]; then
-    brew update || brew update
-    brew outdated pyenv || brew upgrade pyenv
-    brew install pyenv-virtualenv
-    brew install cmake || brew upgrade cmake || true
+    unset PYENV_ROOT
+    curl -L https://github.com/pyenv/pyenv-installer/raw/master/bin/pyenv-installer | bash
+    export PATH="$HOME/.pyenv/bin:$PATH"
 
     if which pyenv > /dev/null; then
         eval "$(pyenv init -)"
@@ -18,7 +17,7 @@ if [[ "$(uname -s)" == 'Darwin' ]]; then
     pyenv activate conan
 fi
 
-pip install conan --upgrade
+pip install conan cmake --upgrade
 pip install conan_package_tools bincrafters_package_tools
 
 conan user